backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
httpx
google-generativeai
solana
python-multipart
redis
python-dotenv
